+++ models/agency/associate.go
@@ -11,15 +11,15 @@ import (
 type UserAgencyAssociate struct {
 	models.BaseModel
 
-	AgencyId uint `gorm:"column:agency_id" json:"agency_id"`
-	UserId   uint `gorm:"column:user_id" json:"user_id"`
+	AgencyId uint64 `gorm:"column:agency_id" json:"agency_id"`
+	UserId   uint64 `gorm:"column:user_id" json:"user_id"`
 }
 
 func (UserAgencyAssociate) TableName() string {
 	return "agency_user_associate"
 }
 
-func GetAssociateById(db *models.MyDB, ctx *gin.Context, id uint) (*UserAgencyAssociate, error) {
+func GetAssociateById(db *models.MyDB, ctx *gin.Context, id uint64) (*UserAgencyAssociate, error) {
 	ret := new(UserAgencyAssociate)
 
 	err := db.Model(&UserAgencyAssociate{}).Where("id=?", id).First(ret).Error
@@ -34,7 +34,7 @@ func GetAssociateById(db *models.MyDB, ctx *gin.Context, id uint) (*UserAgencyAs
 	return ret, nil
 }
 
-func GetAssociateByUserId(db *models.MyDB, ctx *gin.Context, userId uint) (*UserAgencyAssociate, error) {
+func GetAssociateByUserId(db *models.MyDB, ctx *gin.Context, userId uint64) (*UserAgencyAssociate, error) {
 	ret := new(UserAgencyAssociate)
 
 	err := db.Model(&UserAgencyAssociate{}).Where("user_id=?", userId).First(ret).Error
@@ -56,13 +56,13 @@ type AssociateAgencyUserBean struct {
 	Name  string `gorm:"column:name" json:"name"`
 }
 
-func QueryUsersByAgencyId(db *models.MyDB, ctx *gin.Context, agencyId, offset, limit uint) (uint, []*AssociateAgencyUserBean, error) {
+func QueryUsersByAgencyId(db *models.MyDB, ctx *gin.Context, agencyId, offset, limit uint64) (uint64, []*AssociateAgencyUserBean, error) {
 	// conditions
 	tmpDb := db.Table("user").Model(&account.UserBean{})
 	tmpDb = tmpDb.Joins("JOIN agency_user_associate ass ON ass.agency_id = ? AND ass.user_id = user.id AND ass.deleted_at IS NULL", agencyId)
 
 	// 统计总数
-	var total uint = 0
+	var total uint64 = 0
 	err := tmpDb.Count(&total).Error
 	if err != nil {
 		return 0, nil, err
