+++ modules/agency/agencydevice/bean.go
@@ -1,7 +1,7 @@
 package agencydevice
 
 type DeviceBindRequest struct {
-	AgencyId uint   `json:"agency_id"`
-	DeviceId uint   `json:"device_id"`
+	AgencyId uint64 `json:"agency_id"`
+	DeviceId uint64 `json:"device_id"`
 	FileUrl  string `json:"file_url"`
 }
