+++ modules/agency/payment/paymentmethod.go
@@ -21,7 +21,7 @@ func QueryPaymentMethodsHandle(ctx *gin.Context) {
 	}
 
 	ret := &modules.BaseQueryResponse{
-		Total: uint(len(paymentTypes)),
+		Total: uint64(len(paymentTypes)),
 		Data:  paymentTypes,
 	}
